Modified Debian packaging to use 'native' package format.
This will allow aDebian package to be produced directly from the GIT
archive. Otherwise an 'orig.tar.gz' archive is required to build
package without verbose messages (which is a little annoying).

The version number in the debian/changelog file needs to use a
'simple' form. eg. x.y.z